refactor: change initial directory
